Split some functionality into units/core.h
units/core.h: `unit`, dimensional analisys, generic cmath functions, traits (not dimension-specific),
and what they're implemented with (`unit_conversion`, unit manipulators, etc.)

units.h: unit definitions (non-dimensionless), constants, angle cmath functions.
`#include`s the new core header.
